requirements.txt
Django
python-dotenv
transformers
torch
accelerate
tokenizers
pymongo
mongomock